// server/src/types/riotApi.ts
export interface User {
    id: string
    gameName: string
    tagName: string
    puuid: string
    accountId: string
    profileIconId: number
    revisionDate: number
    summonerLevel: number
}

export interface Root {
    metadata: Metadata
    info: Info
  }
  
  export interface Metadata {
    dataVersion: string
    matchId: string
    participants: string[]
  }
  
  export interface Info {
    endOfGameResult: string
    gameCreation: number
    gameDuration: number
    gameEndTimestamp: number
    gameId: number
    gameMode: string
    gameName: string
    gameStartTimestamp: number
    gameType: string
    gameVersion: string
    mapId: number
    participants: Participant[]
    platformId: string
    queueId: number
    teams: Team[]
    tournamentCode: string
  }
// ...
  export interface Team {
    bans: Ban[]
    objectives: Objectives
    teamId: number
    win: boolean
  }
// ...
export interface Participants {
    [key: string]: Participant
}
export interface Game {
    info: Info
    participants: Participants
    teams: Team[]
}
export type Games = Game[]